requirements.txt
Django==1.11.3
argparse==1.2.1
certifi==2017.7.27.1
chardet==3.0.4
djangorestframework==3.6.3
future==0.16.0
idna==2.5
oauthlib==2.0.2
python-twitter==3.3
pytz==2017.2
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
urllib3==1.22
